requirements.txt
aliyun-python-sdk-core-v3==2.13.11
asgiref==3.2.3
cffi==1.14.0
cryptography==2.8
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
django-redis==4.11.0
django-simpleui==3.9.1
djangorestframework==3.11.0
jmespath==0.9.5
Pillow==7.0.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.3
redis==3.4.1
six==1.14.0
sqlparse==0.3.1
